requirements.txt
fastapi
uvicorn
pyyaml
openai==0.28.0
tiktoken
retry
python-dotenv
